david's pr comment changes
